Move user answer draft hook to hooks
